+++ common/src/main/java/com/binaris/wizardry/client/EBClientConstants.java
@@ -1,8 +1,6 @@
 package com.binaris.wizardry.client;
 
 import com.binaris.wizardry.WizardryMainMod;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Style;
 import net.minecraft.resources.ResourceLocation;
 
 public class EBClientConstants {
@@ -22,16 +20,8 @@ public class EBClientConstants {
     public static final int RUNE_TOP = 22;
     public static final int RUNE_WIDTH = 100;
     public static final int RUNE_HEIGHT = 100;
-    public static final int SCROLL_BAR_LEFT = 102;
-    public static final int SCROLL_BAR_TOP = 34;
-    public static final int SCROLL_BAR_WIDTH = 12;
-    public static final int SCROLL_BAR_HEIGHT = 178;
-    public static final int SCROLL_HANDLE_HEIGHT = 15;
     public static final int HALO_DIAMETER = 156;
     public static final int TEXTURE_WIDTH = 512;
     public static final int TEXTURE_HEIGHT = 512;
     public static final int ANIMATION_DURATION = 20;
-    public static final int SEARCH_TOOLTIP_HOVER_TIME = 20;
-    public static final Style TOOLTIP_SYNTAX = Style.EMPTY.withColor(ChatFormatting.YELLOW);
-    public static final Style TOOLTIP_BODY = Style.EMPTY.withColor(ChatFormatting.WHITE);
 }