+++ common/src/main/java/com/binaris/wizardry/content/blockentity/ArcaneWorkbenchBlockEntity.java
@@ -7,7 +7,7 @@ import com.binaris.wizardry.api.content.util.CastItemDataHelper;
 import com.binaris.wizardry.content.item.CrystalItem;
 import com.binaris.wizardry.content.item.SpellBookItem;
 import com.binaris.wizardry.content.menu.ArcaneWorkbenchMenu;
-import com.binaris.wizardry.core.EBConstants;
+import com.binaris.wizardry.core.config.EBServerConfig;
 import com.binaris.wizardry.setup.registries.EBBlockEntities;
 import com.binaris.wizardry.setup.registries.EBItems;
 import com.binaris.wizardry.setup.registries.WandUpgrades;
@@ -51,7 +51,7 @@ public class ArcaneWorkbenchBlockEntity extends BaseContainerBlockEntity {
 
     public static void serverTick(Level level, BlockPos pos, BlockState state, ArcaneWorkbenchBlockEntity entity) {
         ItemStack stack = entity.getItem(ArcaneWorkbenchMenu.CENTRE_SLOT);
-        if (stack.getItem() instanceof IManaItem manaItem && !level.isClientSide && !manaItem.isManaFull(stack) && level.getGameTime() % EBConstants.CONDENSER_TICK_INTERVAL == 0) {
+        if (stack.getItem() instanceof IManaItem manaItem && !level.isClientSide && !manaItem.isManaFull(stack) && level.getGameTime() % EBServerConfig.CONDENSER_TICK_INTERVAL.get() == 0) {
             manaItem.rechargeMana(stack, CastItemDataHelper.getUpgradeLevel(stack, EBItems.CONDENSER_UPGRADE.get()));
         }
     }